fix(api) Organization details should reflect membership
With the settings pages no longer using organization details to get the
team and project lists we can update these properties to reflect actual
membership. This lets us be more efficient for most users as we'll be
loading and serializing only the data they actually use. Once this is
out we can update the UI to skip membership checks as they will no
longer be necessary.

Fixes SEN-64